% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/years.R
\name{suaci_years}
\alias{suaci_years}
\title{Available years for SUACI data}
\usage{
suaci_years()
}
\value{
An integer vector of available years.
}
\description{
Returns an integer vector of the years for which SUACI data is available
on the Buenos Aires Data portal.
}
\examples{
suaci_years()
}
